+++ CoffeeShop/Views/InvoicePage.xaml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Page
+    x:Class="CoffeeShop.Views.InvoicePage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:CoffeeShop.Views"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:user_control="using:CoffeeShop.Views.UserControls.InvoicePage"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="20*"/>
+            <ColumnDefinition Width="80*"/>
+        </Grid.ColumnDefinitions>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*"/>
+        </Grid.RowDefinitions>
+        <user_control:InvoiceListUserControl x:Name="InvoiceListControl"  Grid.Row="0" Grid.Column="0" ItemClick="InvoiceListControl_ItemClick"/>
+        <user_control:InvoiceDetailUserControl x:Name="InvoiceDetailControl"
+                                                Grid.Row="0" Grid.Column="1" />
+    </Grid>
+</Page>
