+++ CoffeeShop/Views/UserControls/HomePage/ChoseDrinkUserControl.xaml
@@ -1,0 +1,127 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="CoffeeShop.Views.UserControls.HomePage.ChoseDrinkUserControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+   xmlns:local="using:CoffeeShop.Views.UserControls.HomePage"
+    xmlns:model="using:CoffeeShop.Models"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:converters="using:CoffeeShop.Helper"
+    mc:Ignorable="d">
+
+    <UserControl.Resources>
+        <converters:TotalPriceConverter x:Key="TotalPriceConverter"/>
+
+
+        <DataTemplate x:Key="ItemOrder" x:DataType="model:DetailInvoice">
+            <Grid Margin="10" Padding="10" CornerRadius="5" Background="{ThemeResource ItemBackground}">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="85*" />
+                    <ColumnDefinition Width="15*" />
+                </Grid.ColumnDefinitions>
+
+                <!-- Item Image and Details -->
+                <StackPanel Grid.Column="0" Orientation="Vertical" Margin="10">
+                    <Grid>
+                        <Grid.ColumnDefinitions>
+                            <ColumnDefinition Width="*"/>
+                            <ColumnDefinition Width="Auto"/>
+                        </Grid.ColumnDefinitions>
+                        <StackPanel Grid.Column="0">
+                            <TextBlock Text="{Binding NameDrink}"  />
+                            <TextBlock Text="{Binding Size}"  Foreground="Gray" FontSize="11"  />
+                        </StackPanel>
+                        <NumberBox Grid.Column="1" x:Name="NumberBoxSpinButtonPlacementExample" HorizontalAlignment="Right"
+                 Value="{x:Bind Quantity, Mode=TwoWay}"
+                 Minimum="1"
+                 SpinButtonPlacementMode="Compact"
+                 SmallChange="1"/>
+                    </Grid>
+                    <TextBox PlaceholderText="Order Note..." FontSize="12" Margin="0,5,0,0" HorizontalAlignment="Stretch" />
+                </StackPanel>
+                <StackPanel Grid.Column="1" HorizontalAlignment="Center">
+                    <TextBlock Text="{Binding Converter={StaticResource TotalPriceConverter}}" Foreground="Gray" FontSize="12" HorizontalTextAlignment="Center" Margin="0,20,0,0"/>
+                    <Button Content="&#xE74D;" FontFamily="Segoe MDL2 Assets"  Background="Transparent" Width="40" Height="32" Margin="0,12,0,0"
+                  Click="TrashButton_Click"  Foreground="{ThemeResource PrimaryText}"/>
+                </StackPanel>
+            </Grid>
+        </DataTemplate>
+    </UserControl.Resources>
+
+    <Grid x:Name="StyledStackPanel2"  Grid.Column="1">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="*"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+            <RowDefinition Height="Auto"></RowDefinition>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="*" />
+        </Grid.ColumnDefinitions>
+
+        <StackPanel  Grid.Column="0" Grid.Row="0"  Padding="10">
+            <TextBlock Margin="10">
+              Order
+            </TextBlock>
+            <StackPanel x:Name="ToggeButtons" Orientation="Horizontal" HorizontalAlignment="Center">
+                <ToggleButton Margin="0,0,20,0"
+              IsChecked="True"
+              Checked="ToggleButton_Checked"
+              Content="1" />
+                <ToggleButton Margin="0,0,20,0"
+              Checked="ToggleButton_Checked"
+              Content="2" />
+                <ToggleButton 
+              Checked="ToggleButton_Checked"
+              Content="3" />
+            </StackPanel>
+        </StackPanel>
+
+        <Grid Grid.Column="0" Grid.Row="1" Margin="10" Padding="10" CornerRadius="5">
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="85*" />
+                <ColumnDefinition Width="15*" />
+            </Grid.ColumnDefinitions>
+
+            <Grid  Margin="10">
+                <Grid.ColumnDefinitions>
+                    <ColumnDefinition Width="*"/>
+                    <ColumnDefinition Width="Auto"/>
+                </Grid.ColumnDefinitions>
+                <TextBlock Text="Name"  FontSize="14" />
+                <TextBlock Grid.Column="1"  HorizontalAlignment="Right" Text="Quantity"  FontSize="14" />
+            </Grid>
+            <TextBlock  Grid.Column="1" Text="Price" Foreground="Gray" FontSize="14" VerticalAlignment="Center" HorizontalAlignment="Center"/>
+        </Grid>
+
+        <Grid  Grid.Column="0" Grid.Row="2">
+            <ScrollViewer VerticalScrollBarVisibility="Auto" HorizontalScrollBarVisibility="Disabled">
+                <ItemsControl ItemsSource="{x:Bind ViewModel.ChosenDrinks}" 
+                            ItemTemplate="{StaticResource ItemOrder}"
+                            HorizontalAlignment="Stretch">
+                    <!--<ItemsControl.ItemTemplate>
+                  </ItemsControl.ItemTemplate>-->
+                </ItemsControl>
+            </ScrollViewer>
+        </Grid>
+        <!-- Add the TextBlock for Total Price -->
+        <StackPanel Grid.Column="0" Grid.Row="3"  Padding="10">
+            <TextBlock Text="Total Price:" FontWeight="Bold" FontSize="16" />
+            <TextBlock Text="{x:Bind ViewModel.TotalPrice, Mode=OneWay}" FontSize="16" Foreground="Gray" />
+        </StackPanel>
+
+        <!-- Add the Button to print content -->
+        <Button Grid.Column="0" Grid.Row="4" Content="Order" Click="OrderButton_Click" HorizontalAlignment="Center" Margin="10"/>
+
+        <ContentDialog x:Name="OrderDetailsDialog"
+         Title="Order Details"
+         CloseButtonText="OK">
+            <StackPanel>
+                <TextBlock x:Name="OrderDetailsTextBlock" Text="Order details" />
+            </StackPanel>
+        </ContentDialog>
+    </Grid>
+
+</UserControl>
