+++ CoffeeShop/Views/UserControls/InvoicePage/InvoiceDetailUserControl.xaml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="CoffeeShop.Views.UserControls.InvoicePage.InvoiceDetailUserControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:CoffeeShop.Views.UserControls.InvoicePage"
+    xmlns:model="using:CoffeeShop.Models"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid>
+        <StackPanel>
+            <TextBlock >
+                 <Run Text="Invoice ID: "/><Run Text="{x:Bind ViewModel._invoice.InvoiceID, Mode=OneWay}"/>
+            </TextBlock>
+            <TextBlock >
+                <Run Text="Customer: "/><Run Text="{x:Bind ViewModel._invoice.CustomerName, Mode=OneWay}"/>
+            </TextBlock>
+            <TextBlock >
+                 <Run Text="Total Amount: "/><Run Text="{x:Bind ViewModel._invoice.TotalAmount, Mode=OneWay}"/>
+            </TextBlock>
+            <ListView x:Name="DetailInvoiceListView" ItemsSource="{x:Bind ViewModel.detailInvoices, Mode=OneWay}" >
+                <ListView.ItemTemplate>
+                    <DataTemplate x:DataType="model:DetailInvoice">
+                        <StackPanel>
+                            <TextBlock Text="{Binding NameDrink}" />
+                            <TextBlock Text="{Binding Quantity}" />
+                            <TextBlock Text="{Binding Price}" />
+                        </StackPanel>
+                    </DataTemplate>
+                </ListView.ItemTemplate>
+            </ListView>
+        </StackPanel>
+    </Grid>
+</UserControl>
