+++ CoffeeShop/Views/UserControls/InvoicePage/InvoiceListUserControl.xaml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="utf-8"?>
+<UserControl
+    x:Class="CoffeeShop.Views.UserControls.InvoicePage.InvoiceListUserControl"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:CoffeeShop.Views.UserControls.InvoicePage"
+    xmlns:model="using:CoffeeShop.Models"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d">
+
+    <Grid>
+        <ListView x:Name="InvoiceListView" SelectionChanged="InvoiceListView_SelectionChanged"
+                  ItemsSource="{x:Bind ViewModel.invoices}" >
+            <ListView.ItemTemplate>
+                <DataTemplate x:DataType="model:Invoice">
+                    <StackPanel>
+                        <TextBlock Text="{Binding InvoiceID}" />
+                        <TextBlock Text="{Binding Status}" />
+                    </StackPanel>
+                </DataTemplate>
+            </ListView.ItemTemplate>
+        </ListView>
+    </Grid>
+</UserControl>
